+++ doc/api_reference.py
@@ -91,6 +91,7 @@ API_REFERENCE = {
                     "SimilarityEncoder",
                     "ToCategorical",
                     "DatetimeEncoder",
+                    "SessionEncoder",
                     "ToDatetime",
                     "ToFloat",
                 ],